pyproject.toml
[project]
name = "drf-docmint"
version = "1.0.0"
description = "A zero-runtime CLI tool for generating static API documentation for Django REST Framework projects."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"



authors = [
    { name = "Adith P", email = "adithprakash008@gmail.com" }
]

keywords = ["django", "drf", "documentation", "api", "openapi", "postman", "pdf", "ast", "static-analysis"]

classifiers = [
    "Environment :: Console",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "markdown>=3.10.2",
    "pyyaml>=6.0.3",
    "rich>=14.3.2",
    "textual>=7.5.0",
    "typer>=0.23.1"
]

[project.optional-dependencies]
pdf = [
    "weasyprint>=68.1"
]

[project.scripts]
doc-mint = "gendoc.main:app"

[project.urls]
Homepage = "https://github.com/adith-p/drf-docmint/"
Issues = "https://github.com/adith-p/drf-docmint/issues"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
